Fix undesired immediate file-opening for artdaq.
artdaq uses the concept of flush principals to signal the end of a
SubRun.  Whenever these flush principals are created it implies to the
state machine (through the Source template) that a Run, SubRun, and
Event must be emitted.  Although writing flush principals is disabled,
opening the output files just before the nop-writeEvent function was
NOT disabled.  The solution is to open the output files only if the
principal in question does NOT correspond to a flush principal.

Making this change in the state-machine Handle* finalize actions
allowed me to remove the flush-id checking from the EventProcessor
functions.