Changes required to resolve a request for an Assns<A, B> with an Assns<A, B, D> if no real Assns<A, B> is available.
Rules:

If there is an Assns<A, B> in the event for a given process it will unambiguously match in preference to any Assns<A, B, D> with the same module label and instance name in the same process.

If there are multiple Assns<A, B, Dn> in the event for a given process (with the same module label / instance name) and no Assns<A, B> for same, then there will be an ambiguity.

If there is an Assns<A, B, D> in one process, and an Assns<A, B> in an earlier process with the same module label / instance name, the former will unambiguously match a request for an Assns<A, B> in preference to the latter unless the process name is specified to match the real Assns<A, B>.

Status:

All existing tests pass. More tests need to be instituted to provide a level of comfort that the system has the intended behavior in as many imaginable scenarios as feasible.

A similar implementation will be necessary for gallery to provide equivalent behavior there.