Encourage more const-correct use of readNext.
Since it is a template, art::Source is somewhat lenient with what the
allowed signature for readNext is.  The following signature are
permissible:

[0] bool readNext(art::RunPrincipal const* const, art::SubRunPrincipal const* const, ...);
[1] bool readNext(art::RunPrincipal const*, art::SubRunPrincipal const*, ...);
[2] bool readNext(art::RunPrincipal* const, art::SubRunPrincipal* const, ...);
[3] bool readNext(art::RunPrincipal* const&, art::SubRunPrincipal* const&, ...);
[4] bool readNext(art::RunPrincipal*, art::SubRunPrincipal*, ...);
[5] other combinations

Of those listed above, only (0) prevents users from calling non-const
member functions of (Sub)RunPrincipal and reassigning the pointer.  We
should be encouraging usage of (0) and not (3), which is what has
formerly been in the documentation.