Adjust semantics of ResultsPrincipal::addOrReplaceGroup (see notes).
The ResultsPrincipal::addOrReplaceGroup was mirrored after the
(Sub)RunPrincipal::addOrReplaceGroup implementation.  However,
whenever the (Sub)RunPrincipal implementation was adjusted for
fragments to accommodate actual replacement of groups, the
ResultsPrincipal implementation was not.

Since Results products are intended to exist outside of the
Run/SubRun/Event-level, the concept of a Results fragment is also
meaningful, since the concept is well-defined in Runs and SubRuns,
which are contained in Results.  To that end, it makes sense for the
group updating to have the same mechanism as the other principals with
RangeSet semantics.