Merge BranchIDListHelper and BranchIDListRegistry.
In the foreseeable future, the BranchIDListRegistry will always be
interacted with in a serialized manner.  For that reason, it should no
longer be an instantiation of art::detail::ThreadSafeIndexedRegistry,
which does explicit locking.

If it is deemed necessary at some point to enable concurrent writing
to the BranchIDListRegistry, all that need happen is the addition of a
static mutex to the class, which should be locked at the beginning of
updateFromInput and updateFromRegistry.