Change to per-branch metadata reading.
In order to support schema evolution of on-disk BranchDescription
objects, the ProductIDStreamer must be initialized with any
BranchIDLists that are on-disk.  Since the BranchIDLists are part of
the same tree as the ProductList (which contains the BranchDescription
objects), it is necessary that the data on the branches must be
retrieved separately.