Switch to using BranchDescription::combinable instead of match.
In the scheme where the BranchDescription::MatchMode is permissive,
which was always the case as far as I could tell, the checks made in
BranchDescription::match are equivalent to those made in
BranchDescription::combinable.  The difference is that whereas 'match'
used the branch name to compare descriptions, 'combinable' uses the
individual branch-name components to check for consistency.