Overhaul FP control to avoid use of unsafe fegetenv().
xpfpa.h is no longer needed.

This service WILL fail on non-Intel platforms until it has been updated to support them.

N.B. INEXACT and DENORM exceptions are disabled in the Intel 387 FPU: this is not currently configurable, but is believed to be the original intent of the code.

N.B. INEXACT and DENORM exceptions are not knowingly disabled in the SSE. Whether to do so should be considered carefully.