Backport MT tests for mock reconstruction workflow.
- Fix Stage 1: event-generation tests
- Fix Stage 2: particle-simulation tests
- Fix Stage 3: reconstruction tests
- Fix Stage 4: concatenation/reordering tests
- Fix Stage 5: analysis tests

The test suite for product aggregation relied on a particular set of
assumptions being met.  The testing has been relaxed so that the final
aggregated products are what are checked, and not intermediary
products.  Checking specific values for intermediate products caused
problems under multi-threaded execution.

This is not a problem with the product-aggregation itself, but rather
in how the tests were structured.