Fix bug #16192: inconsistent provenance for dropped products.
The KeptProvenance class is used for determining what provenance
should be retained, even in the case of dropping products.  The
container that was used for determining which product metadata should
be stored on disk was not in sync with the actual provenance
container.  This introduced a situation where subsequent jobs would
have the product description for the dropped product (via the product
list), but the product provenance entry in the relevant metadata tree
would not include the necessary metadata for the dropped product.

An unconditional dereferencing of a product-provenance ptr (without
checking if the pointer was null) is what led to the segmentation
violation.  Now, an assertion is made that the pointer is non-null
(since it is a reasonable precondition that the provenance pointer be
non-null when retrieving provenance information).