Introduce PerScheduleContainer.
It is a common pattern to replicate data structures across the number
of schedules configured.  The intent of this container is to provide
an std::vector/std::map-like interface for users, but removing the
need for users to explicitly set the size of the container.  This
first commit does not provide the implicit setting of the container
size.