Allow concurrent path execution for modules shared between paths.
One of the concerns is that if a module is shared on more than one
path, a subsequent module on that path could run prematurely.  For
example, consider this configuration:

  p1: [m1, m2]
  p2: [m1, m2]

where m1 produces a data product that is consumed by m2.  With the
current tasking system, both p1 and p2 can execute in parallel.
Module m1 can be executing on path p1, and then path p2 finds out that
since m1 is already executing, it can move on to processing m2.  This
could be a significant problem if m2 requires something produced by
m1.

To solve this problem, a given worker is not allowed to finish
executing its event-level function call until the atomic state of the
worker is no longer 'Working'.  This is an inefficiency if m2 *does
not* use a product from m1.  However, until the consumes information
informs the scheduling, this is the best we can do.

Note that the stalling is *per schedule*.  In other words, there is
nothing that prevents m1 from being executed concurrently on more than
one schedule (i.e. more than one concurrent event).  Because of this,
it is unlikely that the stalling introduced by this commit will
significantly impact any workflows.  The inefficiency should
eventually be reconciled, however.

Note also that for the following configuration:

  p1: [m1, m2]
  p2: [m3, m4]

there is no stalling since there are different workers for each module
specified in the paths.