EndPathExecutor mutex changes.
The ofsMutex_ has been removed since the EventProcessor is ultimately
responsible for querying the state of the output modules (via the
EndPathExecutors)--and all interactions with the end-path executor are
guaranteed to be serialized by the shared end-path executor queue,
owned by the EventProcessor.

I have retained the standard mutex_ in the EndPathExecutor class, but
I've made it a non-static data member.  I changed it to be static
before to guarantee serialized behavior--but the serialization is now
enforced by the shared queue mentioned above.  My guess is that the
EndPathExecutor mutex_ member can go as well, but I won't do that just
yet.