More separation from ROOT.
Change RootMutex to InputSourceMutex.

Remove unnecessary locking calls in RootInputFile (the lock is already
held by the EventProcessor).

The mutex is MixOp should probably be replaced with a mutex specific
to that class--the files associated with mixing operations should not
be related to the ROOT files from the input source.  One should
ascertain whether any of ROOT's global variables requires a close
coordination between the RootInput and MixOp function calls.