Choose different time semantic for file-renaming.
Due to discussion with artdaq, there were some misunderstandings wrt
how file-renaming took place.  The end result of the discussion is
that using the timestamps in the principals is problematic because an
interpretation must be applied to them.

Instead, the file-renamer now uses the times at which the
record(Sub)Run functions are called.