Make ProvenanceDumper more robust against persistence bugs.
art 2.09 introduced a bug where the ROOT branches for dropped products
still written to disk.  This created an undesirable situation where
the framework thought the products were accessible (to some degree),
but they were dummy products that were artifacts of the bug.  To solve
this problem, the determination of product availability was simplified
considerably for the Principal::getForOutput function in art 2.11.

Those simplifications did not persist to the ProvenanceDumper, which
explicitly loops through the Principal groups and does not use the
same logic as the getForOutput function does.  The ProvenanceDumper
logic has been adjusted more closely match getForOutput
implementation.

Resolves bug #21316.