Fix bug #22407--only construct modules present on paths.
This fix includes a breaking change.  Specifically, it is possible for
a user to specify an arbitrary table as a module, as long as that
specification is not included on an execution path (e.g.):

  outputs.BadConfig: {}

is an allowed configuration, as long as the 'BadConfig' module label
does not appear on a trigger/end path.