Reorganization in handling path configurations.
This commit is a major refactorization of steps performed in the
PathManager class' constructor, and the checking that happens during
configuration pruning.

Some of this refactoring is in anticipating of the --config-summary
option that has been requested.  However, duplicate calculations of
path information have been present for some time, and this commit
removes such duplication.